Major improvements to palette import / export: JASC palettes can be imported again, Microsoft palettes are now exported correctly, we are no longer using deprecated methods on NSSavePanel, PXPaletteExporter itself is no longer leaking

+++ Archiving/Palette Exporters/OSPALWriter.m
@@ -13,9 +13,9 @@
 
 typedef struct
 {
-	unsigned long signature;
-	unsigned long fileLength;
-	unsigned long riffType;
+	uint32_t signature;
+	uint32_t fileLength;
+	uint32_t riffType;
 } OSPALHeader;
 
 @implementation OSPALWriter
